refactor(auth): convert if-chain router to declarative route table
Adopts @branch/lambda-http: handler.ts is now a one-line dispatch()
call over routes.ts's ordered Route[] table, keeping the CLI
ROUTES-START/END markers around the table entries.

Moved, not changed:
- controllers/auth.ts: login, respond-challenge, refresh, me, logout
- controllers/register.ts: register, verify-email, resend-code
- controllers/password.ts: forgot-password, reset-password
- services/cognito.ts: cognitoClient, USER_POOL_CLIENT_ID/ID,
  CHALLENGE_SPECS, authResultResponse, challengeResponse,
  mapCognitoAuthError, validatePassword (byte-identical rules)

Local json()/parseBody() deleted in favor of @branch/lambda-http's
versions (identical implementations). Route order matches the
original if-chain exactly. No status codes, response bodies, message
strings, Cognito calls/params, password rules or auth gates changed.

Added @branch/lambda-http as a dependency and regenerated
package-lock.json; extended tsconfig include for controllers/services.
