+++ apps/backend/lambdas/auth/package.json
@@ -27,6 +27,7 @@
   "dependencies": {
     "@aws-sdk/client-cognito-identity-provider": "^3.978.0",
     "@branch/lambda-auth": "file:../../../../shared/lambda-auth",
+    "@branch/lambda-http": "file:../../../../shared/lambda-http",
     "aws-jwt-verify": "^5.1.1",
     "dotenv": "^17.2.3",
     "kysely": "^0.28.10",
